feat(path): Improve Windows path assumptions
Change hard-coded Windows filesystem path assumptions to one that already exists on all
Windows systems (excepting the remaining assumption about the `C:\` drive).  To my mind
this is an improvement, uses fewer features (e.g. `mklink /d ..`), requires less
understanding from the user, requires fewer commands, and is thus both something that
works as is *and* is a better starting place for users to adapt to their specific
circumstances.