feat(debug): Remove hard-coded log path assumption
Mainly I'm removing this to reduce filesystem path assumptions.  AFAICT, we don't need a
log path as we can run the script manually at a `cmd` or `powwrshell` prompt to see
debugging output.

I may yet find reason for a log file as I dig in further and I don't have a strong
opinion about this so if reverting this is a condition of merging, then go ahead.