feat(ifup): Switch to dist network up tools
I think it's better and more maintainable in general to use the same tools and services
the distribution would use to configure and set up the network.  Whatever routing,
nameservers, or other networking configuration the distribution supports will be
supported by this.  For example, on Debian-based distributions, one can use the
`dns-...` options in `/etc/network/interfaces` to configure DNS nameservers in a
standard and recognizable way.