feat(cleanup): More flexible re-use/composition
Improve options for re-use  and composition.  Move site or host  specific details out of
the main `.\resetWindowsNet.ps1` script and into the outer-most batch script intended to
be invoked  by the scheduled  task.  The  outermost script is  much easier for  users to
write themselves and  this be able to  use the main script  without customization.  This
also removes some assumptions  about *other* Hyper-V VMs out of the  main script so that
users that want to use this for WSL2 alone don't get more than they expected.

This does also remove a loop that waits for WSL2 to come up.  I've found this not to be
necessary in all of my testing.  If it is still needed for some use cases, I'm still in
favor of doing it separately, outside the main script.